Array creation

Zarr.zcreate — Function

zcreate(T, dims...;kwargs)

Creates a new empty zarr aray with element type T and array dimensions dims. The following keyword arguments are accepted:

path="" directory name to store a persistent array. If left empty, an in-memory array will be created

name="" name of the zarr array, defaults to the directory name

storagetype determines the storage to use, current options are DirectoryStore or DictStore

chunks=dims size of the individual array chunks, must be a tuple of length length(dims)

fill_value=nothing value to represent missing values

compressor=BloscCompressor() compressor type and properties

attrs=Dict() a dict containing key-value pairs with metadata attributes associated to the array

writeable=true determines if the array is opened in read-only or write mode

Create a new subarray of the group g

Zarr.zzeros — Function

zzeros(T, dims..., )

Creates a zarr array and initializes all values with zero.

Compressors

Zarr.BloscCompressor — Type

BloscCompressor(;blocksize=0, clevel=5, cname="lz4", shuffle=true)

Returns a BloscCompressor struct that can serve as a Zarr array compressor. Keyword arguments are:

clevel=5 the compression level, number between 0 (no compression) and 9 (max compression)

cname="lz4" compressor name, can be one of "blosclz", "lz4", and "lz4hc"

shuffle=true enables/disables bit-shuffling

Zarr.NoCompressor — Type

NoCompressor()

Creates an object that can be passed to ZArray constructors without compression.