API reference

Arrays

Zarr.zcreate — Function

zcreate(T, dims...;kwargs)

Creates a new empty zarr array with element type T and array dimensions dims. The following keyword arguments are accepted:

path="" directory name to store a persistent array. If left empty, an in-memory array will be created

name="" name of the zarr array, defaults to the directory name

storagetype determines the storage to use, current options are DirectoryStore or DictStore

chunks=dims size of the individual array chunks, must be a tuple of length length(dims)

fill_value=nothing value to represent missing values

fill_as_missing=false set to true shall fillvalue s be converted to missings

filters=filters to be applied

compressor=BloscCompressor() compressor type and properties

attrs=Dict() a dict containing key-value pairs with metadata attributes associated to the array

writeable=true determines if the array is opened in read-only or write mode

indent_json=false determines if indents are added to format the json files .zarray and .zattrs. This makes them more readable, but increases file size.

Create a new subarray of the group g

Zarr.zzeros — Function

zzeros(T, dims...; kwargs... )

Creates a zarr array and initializes all values with zero. Accepts the same keyword arguments as zcreate

Group hierarchy

Zarr.zcreate — Method

Create a new subarray of the group g

Zarr.zgroup — Function

zgroup(s::AbstractStore; attrs=Dict())

Create a new zgroup in the store s

Zarr.zgroup — Method

Create a subgroup of the group g

Zarr.zopen — Function

zopen(s::AbstractStore, mode="r"; consolidated = false, path = "", lru = 0)

Opens a zarr Array or Group at Store s. If consolidated is set to "true", Zarr will search for a consolidated metadata field as created by the python zarr consolidate_metadata function. This can substantially speed up metadata parsing of large zarr groups. Setting lru to a value > 0 means that chunks that have been accessed before will be cached and consecutive reads will happen from the cache. Here, lru denotes the number of chunks that remain in memory.

Zarr.zopen — Function

zopen(p::String, mode="r")

Open a zarr Array or group at disc path p.

Zarr.zopen_noerr — Function

zopen_noerr(AbstractStore, mode = "r"; consolidated = false)

Works like zopen with the single difference that no error is thrown when the path or store does not point to a valid zarr array or group, but nothing is returned instead.

Compressors

Zarr.Compressor — Type

abstract type Compressor

The abstract supertype for all Zarr compressors.

Interface

All subtypes of Compressor SHALL implement the following methods:

zcompress(a, c::Compressor): compress the array a using the compressor c.

zuncompress(a, c::Compressor, T): uncompress the array a using the compressor c and return an array of type T.

JSON.lower(c::Compressor): return a JSON representation of the compressor c, which follows the Zarr specification for that compressor.

getCompressor(::Type{<:Compressor}, d::Dict): return a compressor object from a given dictionary d which contains the compressor's parameters according to the Zarr spec.

Subtypes of Compressor MAY also implement the following methods:

zcompress!(compressed, data, c::Compressor): compress the array data using the compressor c and store the result in the array compressed.

zuncompress!(data, compressed, c::Compressor): uncompress the array compressed using the compressor c and store the result in the array data.

Finally, an entry MUST be added to the compressortypes dictionary for each compressor type. This must also follow the Zarr specification's name for that compressor. The name of the compressor is the key, and the value is the compressor type (e.g. BloscCompressor or NoCompressor).

For example, the Blosc compressor is named "blosc" in the Zarr spec, so the entry for BloscCompressor must be added to compressortypes as compressortypes["blosc"] = BloscCompressor.

Zarr.NoCompressor — Type

NoCompressor()

Creates an object that can be passed to ZArray constructors without compression.

Zarr.BloscCompressor — Method

BloscCompressor(;blocksize=0, clevel=5, cname="lz4", shuffle=1)

Returns a BloscCompressor struct that can serve as a Zarr array compressor. Keyword arguments are:

clevel=5 the compression level, number between 0 (no compression) and 9 (max compression)

cname="lz4" compressor name, can be one of "blosclz", "lz4", and "lz4hc"

shuffle=1 Either NOSHUFFLE (0), SHUFFLE (1), BITSHUFFLE (2) or AUTOSHUFFLE (-1). If AUTOSHUFFLE, bit-shuffle will be used for buffers with itemsize 1, and byte-shuffle will be used otherwise. The default is SHUFFLE.

Zarr.ZlibCompressor — Type

ZlibCompressor(clevel=-1)

Returns a ZlibCompressor struct that can serve as a Zarr array compressor. Keyword arguments are:

clevel=-1 the compression level, number between -1 (Default), 0 (no compression) and 9 (max compression)

default is -1 compromise between speed and compression (currently equivalent to level 6).